Added different layout panels and widgets
I have added the method buildUI( ). In this method all the panels,
needed for the layout of UI are created. The widgets do not have any
event handlers yet. There are still some "to do" tasks, like creating
the slider or generating data for the default table presentation. Also
new styles for the flex tabe have to be created in css file. There are
some styles deleted and many added in the css. In html file I changed
the name of the application, but it is not relevant, we can adjust it
anytime