Added possibility to remove filters for city and country
I added remove buttons and ClickHandlers to the city and country
filters. The remove buttons are added to the currentFilterDisplay flex
table in the last column. Clicking on the button calls the onClick
method which removes the filter, being clicked on, and updates the
current filter display, as well as reloads the table.