Changes to the map presentation (year indication under the map + new buttons add map/remove)
I added the year indication to the map (in MapVisualization.java class).
It is visible underneath the map. Futher I deleted the export button and
changed the type of viewMap from HorizontalPanel to FlexTable, so that
all widgets are nicely aligned and potentially one may add a new row
with other data (see ClimateApp.java). I also created a method to
generate the viewMap table, where the map visalization and buttons are
added in the corresponding cells.
TODO: event handlers for the buttons, find out how to create new map
visualization with the same slider and add it as the new row in the flex
table