Adjusted colors for temperatures on the map
A temperature representation a bit more detailed  but less elegant, than
green gradients. May be easily outcommented if we decide to stay with
the green scheme. The problem is that the legend will be always adjusted
to the temperature range each time when a new map is drawn, so this
vivid presentation is not more informative than the green one, one can
not compare the colors only, but still has to look at the temperatures
by hovering over